pyproject.toml
[project]
name = "eet-ragapplication"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.121.0",
    "ipykernel>=7.1.0",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
